### 2023-09-10 #5821 @danielskatz
👋 @range-et - thanks for your submission to JOSS. I'll look for an editor next.

While I do so, you could work on the possibly missing or incorrect DOIs that editorialbot suggests, but note that some suggestions may be wrong. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
