+++ paper.bib
@@ -7,8 +7,7 @@
   pages={105--107},
   year={2015},
   publisher={Nature Publishing Group},
-  doi={10.1038/527105a},
-  url={https://doi.org/10.1038/527105a}
+  doi={https://doi.org/10.1038/d41586-020-00062-z}
 }
 
 @article{tobi2018research,
@@ -43,6 +42,7 @@
   booktitle={European Conference on Object-Oriented Programming},
   pages={257--281},
   year={2014},
+  doi={10.1007/978-1-4842-4979-6_2},
   organization={Springer}
 } 
 
@@ -65,7 +65,8 @@
   number={1},
   pages={7--15},
   year={1989},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1142/9789814434478_0005}
 }
 
 @inproceedings{bourqui2016multilayer,
@@ -85,7 +86,8 @@
   volume={18},
   number={1},
   pages={1--22},
-  year={2017}
+  year={2017},
+  doi={10.21307/joss-2018-003}
 }
 
 @article{granovetter1973strength,
@@ -96,7 +98,8 @@
   number={6},
   pages={1360--1380},
   year={1973},
-  publisher={University of Chicago Press}
+  publisher={University of Chicago Press},
+  doi={10.1093/oso/9780195159509.003.0010}
 }
 
 @misc{visjs,
@@ -110,7 +113,8 @@
   title={Gephi: an open source software for exploring and manipulating networks},
   author={Bastian, Mathieu and Heymann, Sebastien and Jacomy, Mathieu},
   booktitle={International AAAI Conference on Weblogs and Social Media},
-  year={2009}
+  year={2009},
+  doi={10.1609/icwsm.v3i1.13937}
 }
 
 @inproceedings{hagberg2008exploring,
@@ -143,7 +147,8 @@
   journal={Journal of Anthropological Research},
   volume={33},
   pages={452--473},
-  year={1977}
+  year={1977},
+  doi={10.1086/jar.33.4.3629752}
 }
 
 @article{li2021brief,
