+++ paper.bib
@@ -7,7 +7,8 @@
   pages={105--107},
   year={2015},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/d41586-020-00062-z}
+  doi={10.1038/d41586-020-00062-z},
+  url={https://doi.org/10.1038/d41586-020-00062-z}
 }
 
 @article{tobi2018research,
