### 2023-09-10 #5821 @range-et
<!--
/* Font Definitions */
@font-face
	{font-family:"Cambria Math";
	panose-1:2 4 5 3 5 4 6 3 2 4;}
@font-face
	{font-family:Calibri;
	panose-1:2 15 5 2 2 2 4 3 2 4;}
@font-face
	{font-family:"Segoe UI Emoji";
	panose-1:2 11 5 2 4 2 4 2 2 3;}
/* Style Definitions */
p.MsoNormal, li.MsoNormal, div.MsoNormal
	{margin:0in;
	font-size:11.0pt;
	font-family:"Calibri",sans-serif;}
a:link, span.MsoHyperlink
	{mso-style-priority:99;
	color:blue;
	text-decoration:underline;}
code
	{mso-style-priority:99;
	font-family:"Courier New";}
.MsoChpDefault
	{mso-style-type:export-only;}
@page WordSection1
	{size:8.5in 11.0in;
	margin:1.0in 1.0in 1.0in 1.0in;}
div.WordSection1
	{page:WordSection1;}
***@***.*** check references I’ve fixed the DOI references, apologies for any issues. Best,Indrajeet Haldar Sent from Mail for Windows From: Daniel S. KatzSent: Sunday, September 10, 2023 3:30 AMTo: openjournals/joss-reviewsCc: Indrajeet Haldar; MentionSubject: Re: [openjournals/joss-reviews] [PRE REVIEW]: The Plebian Graph Library: A WebGL based network visusalisation and diagnostics package (Issue #5821) 👋 @range-et - thanks for your submission to JOSS. I'll look for an editor next.While I do so, you could work on the possibly missing or incorrect DOIs that editorialbot suggests, but note that some suggestions may be wrong. Please feel free to make changes to your .bib file, then use the command @editorialbot check references to check again, and the command @editorialbot generate pdf when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.—Reply to this email directly, view it on GitHub, or unsubscribe.You are receiving this because you were mentioned.Message ID: ***@***.***>
