### 2023-09-28 #5887 @editorialbot
[REVIEW]: The Plebeian Graph Library: A WebGL based network visusalisation and diagnostics package
**Submitting author:** @range-et (<a href="http://orcid.org/0000-0001-8395-6056">Indrajeet Haldar</a>)
**Repository:** https://github.com/range-et/PGL
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.6
