### 2023-09-29 range-et/PGL#1 @abhishektiwari
Publish PGL as a NPM package
PGL is not distributed as NPM package, that to me seems a very inconvenient way to consume a TypeScript package. Publishign as NPM package is a few hours of work [as described here](https://docs.npmjs.com/creating-and-publishing-scoped-public-packages).

[Joss Review](https://github.com/openjournals/joss-reviews/issues/5887)
