### 2023-09-29 range-et/PGL#2 @abhishektiwari
Add community guidelines
Add clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. [Here](https://github.com/range-et/PGL/community)'s how this project compares to [recommended community standards](https://opensource.guide/).

[Joss Review](https://github.com/openjournals/joss-reviews/issues/5887)
