### 2023-09-29 range-et/PGL#3 @abhishektiwari
Add  automated tests to catch regressions
SDKs/Packages need to set higher bar wrt to quality and tests. How will one catch regressions before releasing a new version? Test can be as simple as taking Screenshot of render of examples included and performing image diff against a refrence image.

[Joss Review](https://github.com/openjournals/joss-reviews/issues/5887)
