### 2023-09-29 #5887 @abhishektiwari
Hi @range-et, Please see my initial comments.

## Functionality

 ### Installation
Package is yet be distributed via NPM. Current way of installation and consuming PGL is not convenient and error-prone. I will strongly recommend that NPM distribution to be enabled before this paper is accepted. Created [GitHub Issue](https://github.com/range-et/PGL/issues/1) on the repository.

After publishing as NPM package, please provide installation instruction for `npm` as well as `yarn` as dependency manager.

In addition, worth calling out what web browsers are supported by PGL (WebGL 1 vs WebGL 2, etc.)

### Performance
Statement of need mentioned performance limitation of SVG based alternatives but does not provide any insight how PGL improves or overcome those concerns. Is the key thesis here is that WebGL is more performant than SVG for rendering large network graphs? If so can we provide appropriate benchmark references comparing SVG vs. WebGL?

### Functionality
Does PGL support both client-side and server-side (headless) WebGL rendering? Unclear from the docs.

## Documentation

### Automated tests
No automated tests found. SDKs need to set higher bar wrt to quality and tests. How will one catch regressions before releasing a new version? Test can be as simple as taking Screenshot of render of examples included and performing image diff against a reference image. Created [GitHub Issue](https://github.com/range-et/PGL/issues/3) on the repository.

### Community guidelines
I was also unable to find explicitly stated community guidelines for anyone wishing to contribute to the software, raise issues, or seek support. Created [GitHub Issue](https://github.com/range-et/PGL/issues/2) on the repository.

## Software Paper

### Quality of writing

#### Spelling/Grammar

`visualisation` not `visusalisation` in title

`JavaScript` not `javascript `

`does the spread appear` not `does the spread appears`

`communication, and documentation` not `communication and documentation`
