+++ test/test.js
@@ -1,0 +1,3 @@
+const pgl = require("pgl");
+
+console.log(pgl)
