+++ test/test.js
@@ -1,3 +1,0 @@
-const pgl = require("pgl");
-
-console.log(pgl)
