### 2024-01-08 bc8e60c
Added in codeof conduct and contributions file
