+++ CONTRIBUTING.md
@@ -1,0 +1,52 @@
+# Contributing to Plebian Graph Library (PGL)
+
+First off, thank you for considering contributing to PGL! It's people like you that make the open source community such a great place. Here are some guidelines to help you get started.
+
+## How to Contribute
+### Reporting Issues
+
+If you find a bug or have a suggestion for a new feature or enhancement, I would love to hear about it! Before you submit an issue, please check to make sure it hasn't already been reported. Here's how you can report an issue:
+
+- Go to the Issues section of our GitHub repository.
+- Click on 'New Issue'.
+- Write a clear title and a detailed description. For bugs, it's helpful to include:
+    * Steps to reproduce
+    * Expected and actual behavior
+    * Screenshots or error messages
+    * Your environment (browser, OS, etc.)
+
+### Submitting Changes
+
+Changes are accepted via Pull Requests (PRs) on GitHub. Here’s how you make a PR:
+
+1. Fork the PGL repository.
+1. Clone your fork to your local machine.
+1. Create a new branch.
+1. Make your changes. Please keep your code as clean and well-organized as possible.
+1. Test your changes.
+1. Commit your changes. Write a clear, descriptive commit message.
+1. Push your changes to your fork on GitHub.
+1. From your fork, make a Pull Request to the PGL main branch.
+1. In your pull request, describe what your changes do and mention any relevant issue numbers.
+
+Please adhere to the coding conventions used throughout the project (indentation, accurate comments, etc.).
+
+### Seeking Support
+
+If you need help or have any questions about contributing, please open an issue in the GitHub Issues. I'd be happy to help!
+
+### Improving Documentation
+
+Good documentation is as important as the code itself. If you find errors, typos, or areas that need enhancement, feel free to create a pull request.
+
+## Code of Conduct
+
+Please follow Code of Conduct in all your interactions with the project.
+
+## Pull Request Process
+
+1. Ensure any install or build dependencies are removed before the end of the layer when doing a build.
+1. Update the README.md with details of changes to the interface, this includes new environment variables, exposed ports, useful file locations, and container parameters.
+1. Increase the version numbers in any examples files and the README.md to the new version that this Pull Request would represent. The versioning scheme used is SemVer.
+1. The Pull Request will be merged once it is approved and passes CI checks. 
+
