+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,40 @@
+# Code of Conduct for Plebian Graph Library (PGL)
+
+## My Pledge
+In the interest of fostering an open and welcoming environment, I, as the maintainer and primary contributor of the Plebian Graph Library (PGL), pledge to make participation in this project a harassment-free experience for everyone, regardless of individual differences.
+
+## Standards
+
+Behavior that contributes to a positive environment includes:
+
+- Using welcoming and inclusive language
+- Showing respect for differing viewpoints and experiences
+- Gracefully accepting constructive criticism
+- Focusing on what is best for the community
+- Demonstrating empathy towards other community members
+
+Unacceptable behavior includes:
+
+- Use of sexualized language or imagery, unwelcome sexual attention or advances
+- Trolling, insulting or derogatory comments, and personal or political attacks
+- Public or private harassment
+- Publishing others' private information without explicit permission
+- Other conduct deemed inappropriate in a professional setting
+
+## Responsibilities
+
+As the project maintainer, I am responsible for clarifying the standards of acceptable behavior and will take appropriate and fair corrective action in response to any instances of unacceptable behavior.
+
+I have the right to remove, edit, or reject comments, commits, code, and other contributions that do not align with this Code of Conduct.
+
+## Scope
+
+This Code of Conduct applies within all project spaces, as well as in public spaces when an individual is representing the project or its community. Representation of the project may include using an official project e-mail address, posting via an official social media account, or acting as an appointed representative at online or offline events.
+
+## Enforcement
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be reported by contacting me [at](indrajeet.haldar08@gmail.com). All complaints will be reviewed and investigated and will result in a response that is necessary and appropriate to the circumstances. Confidentiality will be maintained for the reporter of an incident.
+
+## Attribution
+
+This Code of Conduct is adapted from the Contributor Covenant, version 1.4, available at https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
