### 2024-01-08 #5887 @range-et
Hello! Happy New Year to everyone.

Thank you for being so patient!

I have incorporated a number of changes that @abhishektiwari  had suggested. A few of the major ones being:
1) Adding the library to NPM - although the project is a typescript project the build can be downloaded and used directly, which I agree was unwieldy and hence has been updated and changed.
2) A test to check if the library works - as suggested was added. It does a very basic test of rendering out an example graph to check if any of the dependencies are broken or not. I have intentionally left out some of the more complex visualisation methods since the graph vis field is constantly evolving and more often than not individuals 'layer' visualisation effects as opposed to using one of them.
3) A benchmark test has been performed against D3 which is an industry-standard visualisation library. The details from the benchmark are in the paper and the code to perform the benchmark is available under "benchmark" in the main repository.
4) Other than that a contributions method and a code of conduct has also been added
5) Lastly some edits on the paper wrt to spelling and grammar

Thanks!
