### 2024-01-08 767e498
Updated the documentation and updated the paper
