+++ paper.bib
@@ -158,3 +158,22 @@
   journal={University of Chicago},
   year={2021}
 }
+
+@article{d3ieee,
+author = {Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
+title = {D3 Data-Driven Documents},
+year = {2011},
+issue_date = {December 2011},
+publisher = {IEEE Educational Activities Department},
+address = {USA},
+volume = {17},
+number = {12},
+issn = {1077-2626},
+url = {https://doi.org/10.1109/TVCG.2011.185},
+doi = {10.1109/TVCG.2011.185},
+journal = {IEEE Transactions on Visualization and Computer Graphics},
+month = {dec},
+pages = {2301–2309},
+numpages = {9},
+keywords = {user interfaces, Information visualization, 2D graphics., toolkits}
+}
