+++ paper.md
@@ -1,17 +1,17 @@
 ---
-title: 'The Plebian Graph Library: A WebGL based network visusalisation and diagnostics package'
+title: "The Plebian Graph Library: A WebGL based network visualisation and diagnostics package"
 tags:
-  - Javascript 
+  - JavaScript
   - Visualisation
-  - Graphs 
+  - Graphs
   - Networks
 authors:
   - name: Indrajeet Haldar
     orcid: 0000-0001-8395-6056
-    affiliation: "1" 
+    affiliation: "1"
 affiliations:
- - name: Graduate School of Design, Harvard University, USA
-   index: 1
+  - name: Graduate School of Design, Harvard University, USA
+    index: 1
 date: 8 September 2023
 bibliography: paper.bib
 ---
@@ -20,93 +20,94 @@ bibliography: paper.bib
 
 Given a large network (greater than one million nodes), visualising and diagnosing network data has often proven challenging [@nowogrodzki2015eleven]. Although there is a wide range of statistical tools to draw inferences, the esoteric nature of the statistical analysis of networks limits the communication of the findings to researchers familiar with these research methods [@tobi2018research]. Moreover, the more abstract nature of statistical solutions reduces the complexity of diagnosing network properties, especially network diffusion. It allows for limited documentation and analysis only from the perspective of mathematical models of this particular property of large-scale networks [@gao2012network]. The Plebian Graph Library (PGL) is a library that solves for the visualisation of large networks and their diagnostic study.
 
-The Plebian Graph Library (PGL) is a javascript library (written in Typescript [@bierman2014understanding]) designed to facilitate the visualisation and diagnostic analysis of large-scale network data in browsers using WebGL, using a backend provided by ThreeJS [@danchilla2012three]. Whether dealing with local datasets or data retrieved from online sources (APIs), PGL provides a versatile platform for conducting extensive network simulations, physical modelling, and visualisations whilst offering a range of diagnostic tools for organising network data using standard search algorithms [@mattson2013standards] such as network diffusions, Breadth-first search, Depth-first search and Dijkstra's search algorithm. With a rich set of diagnostic features, including network condensation, weighted edge pruning in highly connected graphs, and support for visualisation techniques like Kamada Kawai layouts [@kamada1989algorithm], Hierarchical plots, Hive plots and Edge bundling [@bourqui2016multilayer], PGL empowers researchers to gain valuable insights from complex network structures. Additionally, PGL contains canonical example of the Zackary's Karate Club (ZKC) dataset [@zachary1977information], and the Erdosh Reyni Random Graph model [@li2021brief] as a generator to study and compare network structures. 
[…]
