### 2024-01-08 range-et/PGL#3 @range-et
Added in a regression test as outlined in the JOSS review
