### 2024-01-08 range-et/PGL#1 @range-et
Now published as an NPM package : https://www.npmjs.com/package/plebiangraphlibrary
Also a boilerplate project is available here: https://github.com/range-et/pgl_example
