### 2024-01-08 range-et/PGL#2 @range-et
Added in community guidelines and a code of conduct
