### 2024-01-18 range-et/PGL#4 @abhishektiwari
Example link is broken
[Examples](https://www.plebiangraphlibrary.com/examples.html) link in readme and [doc portal](https://www.plebiangraphlibrary.com) is broken.

[JOSS Review](https://github.com/openjournals/joss-reviews/issues/5887)
