### 2024-01-18 #5887 @abhishektiwari
@range-et Happy new year. Hope you feeling better.

Thanks for awesome effort on software and taking in all of the feedback.

Please see below,
- [Examples link](https://github.com/range-et/PGL/issues/4) is broken in Readme and documentation portal. Previously it used to work.
- Benchmark should be done side-by-side. Your [benchmarking script](https://github.com/range-et/PGL/blob/main/benchmark/index.js) only has simulation for D3 but not PGL. Maybe I am missing something.
- Lastly, please add Benchmark results to README.md so users can see this upfront.
- I also think `Summary` section is quite long for this paper. Maybe you want move/split the content after paragraph to  a different section like `Introduction` or `Statement of need`
