+++ public/Build/pgl.d.ts
@@ -1,0 +1,642 @@
+import * as THREE from "three";
+import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
+interface Point {
+    x: number;
+    y: number;
+    z: number;
+}
+declare class Point {
+    /**
+     * Constructs a point based on the x y z values
+     * @param x x value
+     * @param y y value
+     * @param z z value
+     */
+    constructor(x: number, y: number, z: number);
+    /**
+     * Displaces a point - note this method moves the existing point
+     * @param Point This is the displacement vactor, used as a point but the same idea holds
+     */
+    translate(Point: Point): void;
+}
+interface Line {
+    points: Point[];
+}
+declare class Line {
+    /**
+     * Constructs a line from an array of points
+     * @param points an array of points
+     */
+    constructor(points: Point[]);
+}
+interface _Node {
+    data: any;
+    neighbours: number[];
+}
+/**
+ * This is the node class - they have an ID which is
+ * essentially an index and some data associated with it
+ * The data also contains the position of the
+ */
+declare class _Node {
+    /**
+     *
+     * @param data Data associated with the node, be sure to be careful to pass in any "pos" data as they correspond to position of the nodes in the visuals of the graph
+     */
+    constructor(data: any);
+}
+interface Edge {
+    start: number;
+    end: number;
+    data: any;
+}
+/**
+ * The edge class, edges have a start and end but they can also have data associated with that edge
+ */
+declare class Edge {
+    /**
+     *
+     * Construct an edge
+     *
+     * @param start Start index of the edge based on the array of nodes
+     * @param end End index of the edge based on the array of nodes
+     * @param data Data associated, note that ldata is reserved for how to draw the lines associated with the edge
+     */
+    constructor(start: number, end: number, data: any);
+}
+export interface Graph {
+    nodes: Map<number, _Node>;
+    edges: Map<number, Edge>;
+}
+/**
+ * The main graph object - this contrains nodes and edges that get modified with different
+ * Operations etc.
+ */
+export class Graph {
+    /**
+     *
+     * Construct a graph object (no initializing)
+     *
+     * @param nodes - Map of all the nodes associated with the graph
+     * @param edges - Map of all the edges assocaiated with the graph
+     */
+    constructor(nodes: Map<number, _Node>, edges: Map<number, Edge>);
+    /**
+     * Prints out a snapshot of data associated with this graph like how many nodes and how many edges
+     */
+    printData(): void;
+    /**
+     *  Initializes the graph and constructs a node adajaceny list list
+     */
+    initialize(): Promise<void>;
+    /**
+     *
+     * This is the official create method to make a graph based on a set of nodes and edges
+     * It also auto initializes the graph and sets all the adjaceny lists in the memory
+     *
+     * @param nodes - map of nodes
+     * @param edges - map of edges
+     * @returns
+     */
+    static create(nodes: Map<number, _Node>, edges: Map<number, Edge>): Promise<Graph>;
+    /**
+     * Constructs the adjacency associated with the graph
+     */
+    constructAdjacencyList(): Promise<void>;
+    /**
+     * Add a noce to the graph
+     * @param nodeID - the node ID
+     * @param data - data associated with the node
+     */
+    add_node(nodeID: number, data: _Node): void;
+    /**
+     * Add an edge to the graph
+     * @param start - Starting index of the edge
+     * @param end - The end index of the edge
+     * @param data - data associated with the edge
+     */
+    add_edge(start: number, end: number, data: any): void;
+    /**
+     *
[…]
