+++ public/Build/pgl.js
@@ -1,0 +1,2960 @@
+var $h9nKb$three = require("three");
+var $h9nKb$threeexamplesjsmlinesLine2js = require("three/examples/jsm/lines/Line2.js");
+var $h9nKb$threeexamplesjsmlinesLineMaterial = require("three/examples/jsm/lines/LineMaterial");
+var $h9nKb$threeexamplesjsmlinesLineGeometry = require("three/examples/jsm/lines/LineGeometry");
+var $h9nKb$threeexamplesjsmcontrolsOrbitControls = require("three/examples/jsm/controls/OrbitControls");
+
+function $parcel$export(e, n, v, s) {
+  Object.defineProperty(e, n, {get: v, set: s, enumerable: true, configurable: true});
+}
+var $parcel$global =
+typeof globalThis !== 'undefined'
+  ? globalThis
+  : typeof self !== 'undefined'
+  ? self
+  : typeof window !== 'undefined'
+  ? window
+  : typeof global !== 'undefined'
+  ? global
+  : {};
+var $parcel$modules = {};
+var $parcel$inits = {};
+
+var parcelRequire = $parcel$global["parcelRequired21f"];
+if (parcelRequire == null) {
+  parcelRequire = function(id) {
+    if (id in $parcel$modules) {
+      return $parcel$modules[id].exports;
+    }
+    if (id in $parcel$inits) {
+      var init = $parcel$inits[id];
+      delete $parcel$inits[id];
+      var module = {id: id, exports: {}};
+      $parcel$modules[id] = module;
+      init.call(module.exports, module, module.exports);
+      return module.exports;
+    }
+    var err = new Error("Cannot find module '" + id + "'");
+    err.code = 'MODULE_NOT_FOUND';
+    throw err;
+  };
+
+  parcelRequire.register = function register(id, init) {
+    $parcel$inits[id] = init;
+  };
+
+  $parcel$global["parcelRequired21f"] = parcelRequire;
+}
+parcelRequire.register("jYcHE", function(module, exports) {
+
+$parcel$export(module.exports, "default", () => $e89d79794dbc0ba9$export$2e2bcd8739ae039);
+
+var $6Xdhg = parcelRequire("6Xdhg");
+
+var $fd3jp = parcelRequire("fd3jp");
+
+var $fw40F = parcelRequire("fw40F");
+
+var $3WMe9 = parcelRequire("3WMe9");
+
+var $cWcXJ = parcelRequire("cWcXJ");
+var __awaiter = undefined && undefined.__awaiter || function(thisArg11, _arguments11, P11, generator11) {
+    function adopt11(value11) {
+        return value11 instanceof P11 ? value11 : new P11(function(resolve11) {
+            resolve11(value11);
+        });
+    }
+    return new (P11 || (P11 = Promise))(function(resolve11, reject11) {
+        function fulfilled11(value11) {
+            try {
+                step11(generator11.next(value11));
+            } catch (e11) {
+                reject11(e11);
+            }
+        }
+        function rejected11(value11) {
+            try {
+                step11(generator11["throw"](value11));
+            } catch (e11) {
+                reject11(e11);
+            }
+        }
+        function step11(result11) {
+            result11.done ? resolve11(result11.value) : adopt11(result11.value).then(fulfilled11, rejected11);
+        }
+        step11((generator11 = generator11.apply(thisArg11, _arguments11 || [])).next());
+    });
+};
+/**
+ * Simulates Kamada kawai for a network in 2d. 3d is not supported yet
+ * Note: This is an async function as it take time for some of the large graphs
+ *
+ * @param Graph - The first input number
+ * @param iterations - The second input number
+ *  @param simulationBound - The bounds of simulation (Mostly a global number to scale the graph up or down)
[…]
