### 2024-02-14 1a3c14a
Updated the spellings in the readme to be american
