+++ paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 Given a large network (greater than one million nodes), visualising and diagnosing network data has often proven challenging [@nowogrodzki2015eleven]. Although there is a wide range of statistical tools to draw inferences, the esoteric nature of the statistical analysis of networks limits the communication of the findings to researchers familiar with these research methods [@tobi2018research]. Moreover, the more abstract nature of statistical solutions reduces the complexity of diagnosing network properties, especially network diffusion. It allows for limited documentation and analysis only from the perspective of mathematical models of this particular property of large-scale networks [@gao2012network]. The plebeian Graph Library (PGL) is a library that solves for the visualisation of large networks and their diagnostic study.
 
-The plebeian Graph Library (PGL) is a JavaScript library (written in Typescript [@bierman2014understanding]) designed to facilitate the visualisation and diagnostic analysis of large-scale network data in browsers using WebGL, using a backend provided by ThreeJS [@danchilla2012three]. Whether dealing with local datasets or data retrieved from online sources (APIs), PGL provides a versatile platform for conducting extensive network simulations, physical modelling, and visualisations whilst offering a range of diagnostic tools for organising network data using standard search algorithms [@mattson2013standards] such as network diffusions, Breadth-first search, Depth-first search and Dijkstra's search algorithm. With a rich set of diagnostic features, including network condensation, weighted edge pruning in highly connected graphs, and support for visualisation techniques like Kamada Kawai layouts [@kamada1989algorithm], Hierarchical plots, Hive plots and Edge bundling [@bourqui2016multilayer], PGL empowers researchers to gain valuable insights from complex network structures. Additionally, PGL contains canonical example of the Zackary's Karate Club (ZKC) dataset [@zachary1977information], and the Erdosh Reyni Random Graph model [@li2021brief] as a generator to study and compare network structures.
+The plebeian Graph Library (PGL) is a JavaScript library (written in Typescript [@bierman2014understanding]) designed to facilitate the visualisation and diagnostic analysis of large-scale network data in browsers using WebGL, using a backend provided by ThreeJS [@danchilla2012three]. Whether dealing with local datasets or data retrieved from online sources (APIs), PGL provides a versatile platform for conducting extensive network simulations, physical modeling, and visualisations whilst offering a range of diagnostic tools for organising network data using standard search algorithms [@mattson2013standards] such as network diffusions, Breadth-first search, Depth-first search and Dijkstra's search algorithm. With a rich set of diagnostic features, including network condensation, weighted edge pruning in highly connected graphs, and support for visualisation techniques like Kamada Kawai layouts [@kamada1989algorithm], Hierarchical plots, Hive plots and Edge bundling [@bourqui2016multilayer], PGL empowers researchers to gain valuable insights from complex network structures. Additionally, PGL contains canonical example of the Zackary's Karate Club (ZKC) dataset [@zachary1977information], and the Erdosh Reyni Random Graph model [@li2021brief] as a generator to study and compare network structures.
 
 An illustrative case for the package is to diagnose large-scale network diffusion. Visualising a clustered network in 3D, where, for example, the network nodes are displaced vertically according to their recursive importance, i.e. eigenvector centralities [@lacobucci2017eigenvector]. A diffusion simulation is then run, and insights and diagnostics of diffusion sequences are gathered. For example, whether diffusion first occurs between high eigenvector centrality nodes across clusters or does the spread appear in the groups before spreading to other clusters, allowing for the visual study of the strength of weak ties behaviour [@granovetter1973strength]. Exploratory research, analysis, communication, and documentation of these network behaviours, as mentioned above, would have been complex in a traditional visualisation library where the emphasis lies on validation instead of exploratory study and diagnostics.
 
