### 2024-03-22 #5887 @majensen
Ok @range-et  - it looks like we can move ahead. I will take a look at the paper itself from an editorial point of view - this may result in some typo fixes or other minor items. Now is the time to create an open archive of the code repo. Please have a look at [this tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) from a previous review. This uses Zenodo, a free service that many JOSS pubs have used. Please report the DOI of the archive and the version back here in this thread. Thanks!
