### 2024-04-07 #5887 @majensen
Hi @range-et - sorry for the delay. Below are some suggestions for making the writing clearer. Please consider these and then let me know when you have made desired changes. Line numbers refer to the latest proof.
---
l. 10: "abstract nature of statistical solutions reduces complexity" -  is the point here that these analyses abstract away complexity that should in fact be included? And that PGL can numerically solve relevant equations in the browser, which enables visualization and simulations that highlight, rather than obscure, inhomogeneities and variability of interest?

l.22, 23, 25, 26 - I would not capitalize the words in algorithm and plot names, except for names of individuals.

l.30 - "Visualising..." -> "We may visualise..." (giving this sentence a subject, otherwise is a fragment.)

l.34 - "For example, whether" -> "For example, we can observe graphically whether...";
l.35 - "or does..." -> "or instead appears in the groups..."
l.36 - "clusters, allowing..." -> "clusters. This enables the visual study..."
l.38 - "complex in" -> "complex using"

l.45 - "D3[d3ieee]": fix reference rendering (or add this ref to bib file?)

l.55 - "involving..." -> "involving rendering a graph of ..." (could "nodes" and "edges" be confused with compute nodes..?)

l.65 - "Since its inception..." - are there other uses besides in the author's thesis? If so, including refs would make this sentence wmore convincing. Alternatively, there is no shame is simply saying that the package played a crucial role in your own research.

l.70 - "dictated" - "strongly influenced"

l.79 - "GitHub for the package" -> "GitHub."
l.79 - "Guides for general guides..." -> "Guides for general usage..." [?]

l.82 - "A boilerplate example...above." - "The example described above is documented at [link]."

l.127 - Reference "V.,A.B."?
