+++ paper.bib
@@ -24,17 +24,13 @@
   url={https://doi.org/10.1007/s11135-017-0513-8}
 }
 
-@article{gao2012network,
-  title={Network model of knowledge diffusion},
-  author={Gao, Xia and Guan, Jiancheng},
-  journal={Scientometrics},
-  volume={90},
-  number={3},
-  pages={749--762},
-  year={2012},
-  publisher={Springer},
-  doi={10.1007/s11192-011-0554-z},
-  url={https://doi.org/10.1007/s11192-011-0554-z}
+@article{Vaishnavi2016A,
+  title={A Visualization Tool for Learning Statistical Analysis in Multi Tabular Datasets},
+  author={K. Vaishnavi and Ashwin Kannan and David Cline and Ronak Etemadpour},
+  journal={2016 IEEE 16th International Conference on Advanced Learning Technologies (ICALT)},
+  year={2016},
+  pages={222-226},
+  doi={10.1109/ICALT.2016.13}
 }
 
 @inproceedings{bierman2014understanding,
@@ -104,10 +100,12 @@
 }
 
 @misc{visjs,
-  title={vis.js},
-  author={Almende B.V.},
-  year={2021},
-  url={https://visjs.org/}
+  title = {{vis.js}: A dynamic, browser-based visualization library},
+  author = {{Almende B.V.}},
+  year = {2017},
+  note = {Available online at: \url{https://visjs.org/}},
+  howpublished = {\url{https://visjs.org/}},
+  license = {Dual licensed under Apache 2.0 and MIT}
 }
 
 @inproceedings{bastian2009gephi,
@@ -159,7 +157,7 @@
   year={2021}
 }
 
-@article{d3ieee,
+@article{D3,
 author = {Bostock, Michael and Ogievetsky, Vadim and Heer, Jeffrey},
 title = {D3 Data-Driven Documents},
 year = {2011},
