### 2024-04-10 #5887 @range-et
> l. 10: "abstract nature of statistical solutions reduces complexity" - is the point here that these analyses abstract away complexity that should in fact be included? And that PGL can numerically solve relevant equations in the browser, which enables visualization and simulations that highlight, rather than obscure, inhomogeneities and variability of interest?

I have reworded this to make it more clear. The point here was that purely statistical analyses of data, even the case of networks often overlooks some nuances that can be verified by a visual check. I have also changed the reference here since I've found a paper that clarifies the point I'm trying to make further

> l.65 - "Since its inception..." - are there other uses besides in the author's thesis? If so, including refs would make this sentence wmore convincing. Alternatively, there is no shame is simply saying that the package played a crucial role in your own research.

I have used it a few times after, but they were in context of on-going research efforts at our lab. Again changed it to reflect the current state of published material, the thesis.

> l.127 - Reference "V.,A.B."? Fixed this reference

Fixed all the other grammatical issues.
