### 2024-04-25 #5887 @danielskatz
I'll check this over again once it's complete...
