+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Haldar
+  given-names: Indrajeet
+  orcid: "https://orcid.org/0000-0001-8395-6056"
+doi: 10.5281/zenodo.10871132
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Haldar
+    given-names: Indrajeet
+    orcid: "https://orcid.org/0000-0001-8395-6056"
+  date-published: 2024-04-25
+  doi: 10.21105/joss.05887
+  issn: 2475-9066
+  issue: 96
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5887
+  title: "The plebeian Graph Library: A WebGL based network
+    visualisation and diagnostics package"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05887"
+  volume: 9
+title: "The plebeian Graph Library: A WebGL based network visualisation
+  and diagnostics package"
