### 2024-04-25 #5887 @danielskatz
Congratulations to @range-et (Indrajeet Haldar) on your publication!!

JOSS depends on volunteers and wouldn't be successful without you
